Create Lab 3
TASKS
1) (2 marks) 
a. Simulate a population of 5,000 with a variable from a Normal(5,4) distribution and a 
population of 5,000 with a variable from an Exponential(1.2) distribution (where 
1.2=lambda). [use the command rexp(n,rate) ] – so rexp(5000, 1.2)
b. Draw samples of size 40 and size 100 from each population.
2) (4 marks)
a. Choose a confidence level and calculate confidence intervals for the mean of the 
sample of size 40 from each distribution, using a) the Z-score and the known 
population SD, and b) the t-distribution and the sample SD.
3) (4 marks)
a. Use R to plot graphs to help you assess the normality of each population, as well as 
each of the samples of size 40. 
b. Would you trust/rely on the confidence intervals you calculated in part 2? Why/Why
not?